Added custom exceptions and FluentVakidation

+++ HRLeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 using HRLeaveManagement.Application.Contracts.Persistence;
+using HRLeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HRLeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -16,6 +17,10 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     {
         // retrieve domain entity object
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+        
+        // verify that record exists
+        if (leaveTypeToDelete == null)
+            throw new NotFoundException(nameof(LeaveType), request.Id);
         
         // remove from database
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
